+++ src/IMUBase.hpp
@@ -50,8 +50,8 @@ public:
 	virtual void getQuat(float* out) = 0;
 	virtual float getTemp() = 0;
 
-	virtual int setGyroRange(uint8_t range) = 0;
-	virtual int setAccelRange(uint8_t range) = 0;
+	virtual int setGyroRange(int range) = 0;
+	virtual int setAccelRange(int range) = 0;
 
 	virtual void calibrateAccelGyro(float* out_accelBias, float* out_gyroBias) = 0;
 	virtual void calibrateMag(float* out_magBias, float* out_magScale) = 0;
